feat(server): validation routes validation routes authorization [#165432620]
